modificacion de admin y formato

+++ protected/controllers/ProfesorguiacpController.php
@@ -190,8 +190,9 @@ class ProfesorguiacpController extends Controller
 	public function actionDelete($id)
 	{
         $studentData=Estudiante::model()->find('ProfesorGuiaCP_RutProfGuiaCP=?',array($id));
+        $planningData=Planificacionclaseadministrador::model()->find('ProfesorGuiaCP_RutProfGuiaCP=?',array($id));
 		
-		if($studentData == null){
+		if($studentData == null && $planningData == null){
             $this->loadModel($id)->delete();
             
             // if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
@@ -199,7 +200,7 @@ class ProfesorguiacpController extends Controller
                 $this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
         
         }else{
-            Yii::app()->user->setFlash('message',"<div id='errorMessage' class='flash-error'><p><strong>¡No es posible eliminar!</strong></p><ul><li>Hay estudiantes asociados a profesor cp.</li></ul></div>");
+            Yii::app()->user->setFlash('message',"<div id='errorMessage' class='flash-error'><p><strong>¡No es posible eliminar!</strong></p><ul><li>Hay estudiantes y/o planificaciones asociados a profesor cp.</li></ul></div>");
 			//$this->refresh();
 			$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array("view&id=".$id.""));
 		}
